Fix the limits enforcement of message stores
File message stores use multiple files to store messages on a given
channel. The idea is that the amount of messages is spread accross
multiple files, but allows for 25% more messages. When reaching
125%, the first file is removed, and files are shifted.
The problem was that the user should still see strict enforcement
of the limits, in otherwords, when quering the stats of a message
store, we should never get more than 100% of the limit.

Adding also a different test default datastore per package to
avoid clashes during parallel tests.

Tweaked a test to reduce flapping on Windows.

Resolves #39