[CHANGED] Option parameter to connect to external NATS Server
You can now use `-nats_server <url>` instead. When you provide
the url to connect to a NATS server that requires authentication,
you can pass the username and password in the url itself, or
with `-user` and `-pass`. If username and password is present
in the url, they will override the one passed through the
command line.

This PR is related to #125