[IMPROVED] Performance by selectively calling Debugf/Tracef
Debugf/Tracef check an atomic variable to see if the statement
should be logged or not. However, the cost of calling these
functions is significant due to memory allocations caused by
the formatting of the string being logged.
In places where speed matters, this is too much. Using server's
boolean (that capture the configured debug/trace options) and use
those to surround Debugf/Tracef calls improve performance by
limiting the calls to only when debug and/or trace are actually
enabled.