Allow "unlimited" channel limits
On startup, the server creates a store with default store limits.
In order to override those default limits, you can specify them
through Options or command line parameters. To allow a metric to
be ignored, the value -1 (or anything negative) is used to indicate
that the given limit should not be checked.
For instance, you could start a server configuring the max number
of messages, but not limiting the size. You would start the server
like this:

nats-streaming-server --max_msgs 10000 --max_bytes -1