[ADDED] Server now logs the cause of a start failure.
The server panics if it fails to startup. The problem is that if
logging is used and oen has only access to the log (not the console),
one would not see the cause of the failure to start.
This change ensures that the cause of the panic is logged.
The server still panics on failed start.

Resolves #173