[ADDED] File slices configuration, including script for archiving
You can now configure file slices to be of a certain size. The channel
limits may still override the size of the slices.
When a file slice is supposed to be removed, if a script has been
defined, the script is invoked with the renamed file slice. The
script can then move those files for archiving.

Related to #192