[FIXED] Expiration timer runaway when some but not all messages expire
When the expiration timer fires, messages that have expired are
removed. The timer is then reset to the first message yet to expire
based on the message’s time left to reach the target expiration
time. A bug was reseting the timer to a negative value, which Go
would translate to setting the timer to a nano-second.