[FIXED] Client's lock held for ClientHBTimeout in checkClientHealth
The timer sending HB requests to the client was holding the client
lock for the duration of the request. Since the HBTimeout can
be quite large, this was problematic since it could block routines
in the server trying to lock the client object if the client was
disconnected and the timer was waiting for the HB reply.

Resolves #234