[FIXED] Sending data before SubscriptionResponse
When a subscription request is received, the subscription is added
to the server and the response is sent back to the client before
starting to deliver messages to this subscription. However, in
the subscription request process, as soon as the subscription is
added but before the response is sent, if an application is sending
messages to the channel, it was possible that server would start
sending data to that subscription that had not yet fully initialized.

The issue was discovered with the node client:
https://github.com/nats-io/node-nats-streaming/issues/31

This was not affecting other Streaming client due to the use of
locking when processing the subscription request and processing
of incoming data.