[ADDED] Shared Storage Fault Tolerance
Implementation of Fault Tolerance with shared storage.
In a given group of serveral FT servers, only 1 will act as the
active server (accessing the storage and servicing clients) while
the others will act as standby waiting for the active failure.
They communicate through heartbeats, but the decision to become
active relies on the ability to a server to get an exclusive
lock on the shared storage.

Resolves #263