Changed error texts in fmt.Errorf() calls to not use uppercase
The rationale is that when a function returns an error, that function
does not know if the error is going to be printed/logged as standalone
or as a suffix of another error message. So functions should not
return error texts with uppercase.

First letter changed to lower case except for:
- tests (avoid using uppercase in the future, but no need to go
  over all tests to change them now)
- panics: since this is the only error that is going to be printed,
  starting with uppercase is ok