Updated Store's dropping messages notice to use friendly bytes
When a channel hits a limit, the store prints a message with
the current/limit values for number of messages and bytes.
This PR prints the size as KB, MB, etc.. instead of raw numbers.

To do so, I have promoted the server's friendlyBytes function
to a util function, and added tests for that.