[ADDED] Partitioning
This feature allows to have multiple server instances on the
same cluster ID but each instance handling a list of channels.
A given channel must not be defined on more than one instance.
Servers on the same cluster ID try to ensure that this is the case.

This feature also allows to prevent creation of unwanted channels.
A similar feature was proposed by @maxatome in PR #265 that was
not accepted because there were plans for FT and Channels Partitioning.

Resolves #271