[FIXED] Unable to set a per-channel limit if global limit set to 0
This was introduced in v0.4.0 when fixing an issue where server
would not report negative values. Basically, prior to v0.4.0, the
limit check preventing a per-channel to be lower than the global
limit was not enforced (but was meant to be).
In v0.4.0 the rule is enforced and with this fix it will be possible
to set a global limit to 0 (unlimited) and lower the limit of a
specific channel.

In future release, we are going to remvoe the check that enforces
a channel limit to be lower than the corresponding global limit.

Resolves #292