[ADDED] Ability to combine NATS and Streaming options in same config
NATS and NATS Streaming options do not overlap, so they can be
combine in the same configuration file. This was already the case
but users would have to pass the unified configuration file with
both `-sc` and `-c` command line parameters.
This PR simply makes the server tries to parse the configuration
file given by `-sc` or `-c` to both NATS and Streaming components.

Resolves #311