[FIXED] Server panics when more than one shadow queue sub is recovered
A shadow queue subscriber is a record left in the store when the last
member of a queue group has left, but the state needs to be maintained.
There should be only one such shadow queue durable subscriber per
queue group per channel.
However, we got a case with a datastore that contained 2 such records.
Code inspection did not reveal how this could happen with current
code base, so it is suspected that this is the result of various
server upgrades with older datastores.
This change prevents the server from panic'ing and instead uses
the shadow subscriber with the highest lastSent value.

Resolves #322