Fixed timers and flapping test
Timer.Reset() should not be called concurrently with a timer
actually expiring. In one of the test, the timer's callback was
invoked directly, with a call to Reset() inside that callback
while the timer may have been firing. This caused the call to
Timer.Stop() to return `true` and yet have the callback firing,
which would cause a panic when calling WaitGroup.Done().

Tweaked a flapping test.