Fixed logging time value when config file is specified
We default the logging of time as true, which is a problem when
calling natsd.ProcessConfigFile() since this is returning a new
natsd.Options and so it is impossible to apply the default value
without possible overriding the value that would be set in the
file (that is, if explicitly set to false).
This is hacky: a proper fix would be to add a NATS ProcessConfigFile
version that would accept an Options and fill it out with what is
in the file. This would preserve whatever default we set in the
option given to this new function.