Fixed SubStore handling of subscriptions count
The `genericSubStore` was keep track of the subscriptions count
using a simple integer. The problem is that on DeleteSub(), the
count would be decreased, regardless if this subscription either
existed or had been already deleted or never existed.
This was not really a problem since the server should not delete
the same subscription twice.
However, it is safer to make the DeleteSub() store operation
idempotent. To that effect, the count is replaced with a map
in the genericSubStore. The test for max number of subscriptions
now relies on the length of the map. If a subscription is deleted
twice, it will be removed from the map only once and again, the
count is based on size of map, not a separate counter.
Since FileSubStore used a map due to the need during compaction
of the subscriptions file, the field has been removed and instead
the FileSubStore is now using the map from the genericSubStore.
That map uses interface{} for the values. The FileSubStore therefore
type cast the value to (*subscription).