[FIXED] After restart, server sends msgs to offline durables
After the server has recovered subscriptions, it detects durables
that are offline (this is a durable subscription for which client
connection was not recovered). However, they were still added
to the list of subscriptions, which means that when a message
was received on a durable's subject, the server will attempt to
deliver it to the durable's old inbox. This would also cause the
message to be added to the pending map, which means that they
would be redelivered every AckWait seconds.
We ensure now that a durable is added back to the list of active
subscriptions only if it is not offline.

Resolves #380