Fixed stalled queue
This was introduced by #378. The queue group as a whole may not
be stalled (that is, not all members are stalled), but if a server
processes the ack of one of the member that causes that member
to unstall, then it needs to attempt to send available messages.
It was not doing that.

Related to #378