Allow configuring fsync for Raft log
Apply the fsync configuration from the file store to the Raft log
backend. This allows disabling disk syncing on every write to the
Raft log which improves performance. While this creates some risk,
the thought is if we are replicating to enough nodes, the replication
itself is sufficient for HA of data since the likelihood of more than a
quorum of nodes failing is low.