Need to update LastSent when ack'ing last seq of a subscription
In SQL store implementation, ack'ing a message deletes the corresponding
row in SubsPending table. So on recovery, we would lose the knowledge
of what was the last sent message.
Added a `lastsent` column in the Subscriptions table. The row is
updated whenever AckSeqPending() is called with the highest seq
for this subscription. This allows to keep track of the LastSent
in the recovery process.