Fix clustered channel recovery
If there are multiple channels, recovering a cluster could deadlock if
the servers recover the channels in different orders. This is because we
wait for a Raft leader to be elected and if we don't elect a leader in
30 seconds, we panic. To fix this, only wait for a leader when creating
the channel "online". If we're recovering, don't wait for the leader.