[ADDED] Support for FT with SQL Store
Simulate locking of database with use of table with single record
that is updated with owner (nuid) and tick. The table is updated
at regular interval. If owner is empty, lock can be acquired right
away. If not, instance checks if tick is updated after waiting for
1.5*interval. If tick has changed, lock is assumed healthy.
If after a configurable amount of non updates, an instance can
steal the lock.
If the owner of the lock fails to update the lock table, the
process will panic.