[CHANGED] FT Standby exit on error trying to get exclusive lock
An FT Standby server that tried to get the exclusive lock when
missing FT heartbeats would exit if it got an error (as opposed
to knowing that it did not get the lock). In otherwords, the error
was considered fatal.
But with the introduction of SQL store, losing the DB connection
would result in standby exiting. So the server will now report
the error but go back to the loop of trying again after the
missed HB interval.

Resolves #446