[FIXED] Remove ghost queue durable when such record is recovered
In issue #215, we incorrectly kept the record of a queue durable
member that closed. After the fix, on recovery this record would
not be recovered and a warning was printed. However, the record
was not deleted and subsequent restarts would continue to print
the warning.

This PR will print the warning as usual but then delete the record,
which means that at the next restart, the record (and therefore
warning) should no longer be recovered.