[FIXED] Clustering: should not start if streaming but no raft state
The raft state is persisted in a directory that is configurable
with `--cluster_log_path`, but otherwise defaults to the name
of the cluster.
If after restarting a node the raft store has been cleared (as it
could happen when you misconfigure a node using Docker by having
the raft store being created in the container and therefore not
surviving a restart), bad things can happen: a node could elect
itself leader although it should not, or be unable to rejoin the
cluster. Also, the discrepancy between the streaming store and
raft store could cause other issues such as panic dealing with
the message store.

Resolves #497