[IMPROVED] Performance after Raft log snapshot
When Raft deletes a lot of entries in the raft log (backed by BoltDB),
the performance of subsequent writes would suffer.
Added some logic to possibly recreate new log with remaining entries
if that number is within some arbitrary limits.
Added logging to report compaction duration, and if that takes more
than 2 seconds, warn the user that the TrailingLogs value should
probably be lowered.

Resolves #520