[FIXED] Avoid possible duplicate redeliveries after message is ack'ed
If the streaming server is connecting to a standalone NATS Server,
and the streaming server is redelivering messages, if a disconnect
occurs, these redeliveries may be buffered by the NATS client
library used by the Streaming server. When the NATS Server is restarted
and the Streaming server reconnects, the pending buffer is flushed,
which if the client has reconnected before the server, would cause
the client to receive possibly a serie of the same redelivered
messages, even if the client is ack'ing them.
To prevent this, set the reconnect buffer to -1 to prevent any
buffering.