[ADDED] Support for client pings
Server can now receive client PINGs and not only echo back, but
can send a response indicating if the client sending the PING
is still valid or not. Indeed, a client could be "separated"
from the server and another client with the same client ID could
start and the old client be replaced. When the old client communication
with the server is restored, there would be a situation where two
clients with the same ID would exist.
The old client would now know from the PING response that it is
no longer valid.
Also, published message now contain a client UID that allows the
server to reject messages from the old client in the situation
described above.

This PR contains changes on the server. Clients implementations
to follow.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>