Update an SQL statement for Postgres
This is related to #569.
It looks like the SQL statement for postgres driver against a
cockroach db would not work. It works fine with Postgres database.

I had another look at that specific statement and rewrote it for
both MySQL and Postgres. This statement works now against a
cockroach db.. However, there would still be an issue in using
cockroach db it seems due to the "SELECT FOR UPDATE" for locking
that seem is not supported there.

Still, there may be value in updating this statement.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>