[FIXED] Durable subscriptions miss messages after cluster restart
If a durable subscription (can apply to durable queues too) is
created with a start position other than SequenceStart, and this
subscription receive some messages, then is closed. At that time
either more messages are sent or there were already more messages
in the channel, then after a cluster restart (all nodes are
restarted), then when the durable (queue) is resumed, it would
seem to not receive any messages in the channel. Only new messages
produced after the cluster is restarted would be delivered.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>