[FIXED] After server restart client sending PINGs are disconnected
This is for newer clients sending PINGs to the server. If a server
is restarted, the information allowing the server to detect if
the client is still valid was not properly recovered, causing the
server to return an error back to the client which then caused it
to disconnect.

This does not affect older clients or when servers were running
in clustering mode.

Resolves #592

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>