Added Close() for StanLogger to allow resources to be released
If the logger is file based, allow file to be closed for proper
cleanup. This was found when running tests on Windows where test
log could not be cleaned-up because the file was still opened.

Also added some missing defers that would also cause some files
to fail to be deleted on test exit.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>