[FIXED] SQLStore: Incorrect first/last seq after recovery and expired msgs
There could be an issue with the SQLStore when server is restarted
and messages are expired during the recovery process. The first/last
sequence for a channel could be incorrect which would cause subscriptions
to possibly not receive messages since their last_sent could be higher
than the new last_seq for that channel.

Resolves #655

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>